+++ tests/container/mock_pkg/device/mock_motors.py
@@ -11,7 +11,6 @@ class MyMotor(StandardReadable):
     def __init__(
         self,
         name: str,
-        /,
         *,
         egu: str = "mm",
         integer: int = 0,